requirements.txt
# Web App UI
streamlit==1.33.0

# OpenAI LLMs
openai==1.30.1

# Document Parsing
PyMuPDF==1.24.2  # for reading PDFs
pdfminer.six==20221105  # optional fallback for PDF parsing

# LangChain & Embeddings
langchain==0.2.1
faiss-cpu==1.7.4
tiktoken==0.7.0
sentence-transformers==2.6.1  # optional for custom embeddings
chromadb==0.4.24  # optional if using Chroma DB

# Utils
python-dotenv==1.0.1  # if using .env for API keys

# LLMs (if needed for future expansion)
transformers==4.42.2
faiss-cpu
openai
streamlit
python-dotenv
PyMuPDF
langchain
langchain-community
tiktoken
